# public/openapi.yaml
openapi: "3.0.3"
info:
  title: 智学AI - 开放API
  description: |
    智学AI教育平台开放API接口文档。提供AI辅导、LLM代理、题库管理、游戏化学习等接口。

    ## 快速开始
    1. **注册账号** - 在智学AI平台注册账号
    2. **获取API Key** - 在个人中心的「API Keys」页面创建虚拟API Key
    3. **调用API** - 使用 `X-API-Key` 请求头携带API Key发起请求

    ## 认证方式
    - **API Key** (`ApiKeyAuth`): 通过 `X-API-Key` 请求头或 `Authorization: Bearer <key>` 传递
    - **会话认证** (`BearerAuth`): 通过 NextAuth 的 session cookie 自动认证（仅限平台内调用）

    ## 错误响应
    所有接口在发生错误时都会返回统一格式：
    ```json
    { "error": "错误描述信息" }
    ```

    常见状态码：
    - `400` - 请求参数错误
    - `401` - 未认证 / API Key无效
    - `402` - 余额不足
    - `403` - 无权限
    - `404` - 资源不存在
    - `429` - 请求频率超限
    - `500` - 服务器内部错误

    ## 速率限制
    - 免费用户: 50 RPM
    - VIP用户: 200 RPM
    超出限制后返回 `429` 状态码。

  version: "1.0.0"
  contact:
    name: 智学AI开发团队
    email: dev@zixueai.com
    url: https://zixueai.com
servers:
  - url: "{baseUrl}"
    description: 智学AI API服务器
    variables:
      baseUrl:
        default: "https://api.zixueai.com"
        description: API服务器地址
        enum:
          - "https://api.zixueai.com"
          - "http://localhost:3000"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: "从个人中心获取的API Key，格式: sk-edu-xxxxxx-xxxxxxxx"
    BearerAuth:
      type: http
      scheme: bearer
      description: "NextAuth Session Token（平台内自动携带）"
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: 错误描述信息
          example: "请先登录"
      required:
        - error
...
    RechargeRequest:
      type: object
      properties:
        amount:
          type: number
          description: 充值金额（CNY），范围 1-10000
          example: 10
          minimum: 1
          maximum: 10000
        paymentMethod:
          type: string
          enum: [alipay, wechat]
          description: 支付方式
          example: "alipay"
      required:
        - amount
        - paymentMethod
    RechargeResponse:
      type: object
      properties:
        id:
          type: string
          example: "rech_abc123"
        amount:
          type: number
          example: 10
        paymentMethod:
          type: string
          example: "alipay"
        status:
          type: string
          enum: [pending, success, failed]
          example: "pending"
        paymentUrl:
          type: string
          description: 支付链接（生产环境返回）
          example: "https://pay.example.com/order/abc123"
        balance:
          type: number
          description: 充值后的余额（仅开发环境自动完成时返回）
          example: 22.50
        createdAt:
          type: string
          format: date-time
...
    Question:
      type: object
      properties:
        id:
          type: string
          example: "q_abc123"
        subject:
          type: string
          example: "数学"
        type:
          type: string
          enum: [choice, fill, essay]
          example: "choice"
        difficulty:
          type: integer
          minimum: 1
          maximum: 5
          example: 3
        content:
          type: string
          example: "在直角三角形ABC中..."
        options:
          type: array
          items:
            type: string
          example: ["A. 勾股定理", "B. 相似三角形", "C. 面积公式", "D. 三角函数"]
        answer:
          type: string
          example: "A"
        explanation:
          type: string
          example: "勾股定理是直角三角形的基本性质..."
        source:
          type: string
          example: "2024年中考真题"
        gradeLevel:
          type: string
          example: "初三"
        knowledgePoints:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              name:
                type: string
              subject:
                type: string
        errorNoteCount:
          type: integer
          example: 5
        createdAt:
          type: string
          format: date-time
...
    ErrorNote:
      type: object
      properties:
        id:
          type: string
          example: "en_abc123"
        question:
          $ref: "#/components/schemas/Question"
        wrongAnswer:
          type: string
          example: "C"
        correctAnswer:
          type: string
          example: "B"
        status:
          type: string
          enum: [active, eliminated]
          example: "active"
        timesWrong:
          type: integer
          example: 3
        createdAt:
          type: string
          format: date-time
        lastMistakeAt:
          type: string
          format: date-time
    ErrorNoteListResponse:
      type: object
      properties:
        errorNotes:
          type: array
          items:
            $ref: "#/components/schemas/ErrorNote"
        total:
          type: integer
          example: 45
        totalPages:
          type: integer
          example: 3
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
...
paths:
...
  /api/error-notes:
    get:
      tags:
        - Questions
      summary: 获取错题本
      description: 获取当前用户的错题本列表，支持按学科和状态筛选
      operationId: getErrorNotes
      security:
        - BearerAuth: []
      parameters:
        - name: subject
          in: query
          description: 学科筛选（为空则全部）
          required: false
          schema:
            type: string
          example: "数学"
        - name: status
          in: query
          description: 错题状态筛选
          required: false
          schema:
            type: string
            enum: [active, eliminated]
        - name: page
          in: query
          description: 页码（从1开始）
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: 每页数量（最大50）
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
      responses:
        "200":
          description: 错题列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorNoteListResponse"
              example:
                errorNotes:
                  - id: "en_abc123"
                    question:
                      id: "q_abc123"
                      subject: "数学"
                      type: "choice"
                      difficulty: 3
                      content: "已知直角三角形两条直角边分别为3和4，则斜边长为？"
                      options: ["A. 5", "B. 6", "C. 7", "D. 8"]
                      answer: "A"
                      explanation: "根据勾股定理：3²+4²=9+16=25"
                    wrongAnswer: "C"
                    correctAnswer: "A"
                    status: "active"
                    timesWrong: 3
                    createdAt: "2026-05-07T10:00:00Z"
                    lastMistakeAt: "2026-05-07T10:05:00Z"
                total: 45
                totalPages: 3
                page: 1
                limit: 20
        "401":
          description: 未登录
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ==================== Game ====================
...
  /api/billing/recharge:
    post:
      tags:
        - Billing
      summary: 账户充值
      description: |
        创建充值订单。
        - 开发环境：自动完成充值
        - 生产环境：生成支付链接
      operationId: recharge
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RechargeRequest"
            example:
              amount: 10
              paymentMethod: "alipay"
      responses:
        "200":
          description: 充值结果
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RechargeResponse"
              examples:
                dev:
                  summary: 开发环境（自动完成）
                  value:
                    id: "rech_abc123"
                    amount: 10
                    paymentMethod: "alipay"
                    status: "success"
                    balance: 22.50
                    createdAt: "2026-05-07T10:00:00Z"
                production:
                  summary: 生产环境（待支付）
                  value:
                    id: "rech_abc123"
                    amount: 10
                    paymentMethod: "alipay"
                    status: "pending"
                    paymentUrl: "https://pay.example.com/order/abc123"
                    createdAt: "2026-05-07T10:00:00Z"
        "400":
          description: 参数错误
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "充值金额必须为有效数字"
        "401":
          description: 未登录
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...